Improve report
